requirements.txt
Flask==3.0.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
# SQLite is built into Python, no additional package needed
